fix 金额转换错误导致的无法回写订单表数据库

+++ src/main/java/net/bus/web/model/Pojo/AlipayAsyncCallBack.java
@@ -120,6 +120,6 @@ public class AlipayAsyncCallBack extends AsyncCallBack{
 
     @Override
     public BigDecimal getPay(){
-        return BigDecimal.valueOf(Long.valueOf(getAmount()));
+        return BigDecimal.valueOf(Double.valueOf(getAmount()));
     }
 }
